+++ scrapy_poet/downloadermiddlewares.py
@@ -148,9 +148,11 @@ class InjectionMiddleware:
         )
         # Fill the callback arguments with the created instances
         for arg, value in final_kwargs.items():
-            # Precedence of user callback arguments
-            if arg not in request.cb_kwargs:
-                request.cb_kwargs[arg] = value
+            # If scrapy-poet can't provided the dependency, allow the user to
+            # give it.
+            if value is None and arg in request.cb_kwargs:
+                continue
+            request.cb_kwargs[arg] = value
             # TODO: check if all arguments are fulfilled somehow?
 
         return response