+++ tests/test_retries.py
@@ -1,8 +1,10 @@
 from collections import deque
+from typing import Any, List
 
 from pytest_twisted import inlineCallbacks
 from scrapy import Request, Spider
 from web_poet.exceptions import Retry
+from web_poet.page_inputs.http import HttpResponse
 from web_poet.pages import WebPage
 
 from tests.utils import EchoResource, MockServer, make_crawler
@@ -21,15 +23,21 @@ class BaseSpider(Spider):
     }
 
 
+def _assert_all_unique_instances(instances: List[Any]):
+    assert len({id(instance) for instance in instances}) == len(instances)
+
+
 @inlineCallbacks
 def test_retry_once():
     retries = deque([True, False])
-    items = []
+    items, page_instances, page_response_instances = [], [], []
 
     with MockServer(EchoResource) as server:
 
-        class ItemPage(WebPage):
+        class SamplePage(WebPage):
             def to_item(self):
+                page_instances.append(self)
+                page_response_instances.append(self.response)
                 if retries.popleft():
                     raise Retry
                 return {"foo": "bar"}
@@ -38,7 +46,7 @@ def test_retry_once():
             def start_requests(self):
                 yield Request(server.root_url, callback=self.parse)
 
-            def parse(self, response, page: ItemPage):
+            def parse(self, response, page: SamplePage):
                 items.append(page.to_item())
 
         crawler = make_crawler(TestSpider, {})
@@ -49,18 +57,22 @@ def test_retry_once():
     assert crawler.stats.get_value("retry/count") == 1
     assert crawler.stats.get_value("retry/reason_count/page_object_retry") == 1
     assert crawler.stats.get_value("retry/max_reached") is None
+    _assert_all_unique_instances(page_instances)
+    _assert_all_unique_instances(page_response_instances)
 
 
 @inlineCallbacks
 def test_retry_max():
     # The default value of the RETRY_TIMES Scrapy setting is 2.
     retries = deque([True, True, False])
-    items = []
+    items, page_instances, page_response_instances = [], [], []
 
     with MockServer(EchoResource) as server:
 
-        class ItemPage(WebPage):
+        class SamplePage(WebPage):
             def to_item(self):
+                page_instances.append(self)
+                page_response_instances.append(self.response)
                 if retries.popleft():
                     raise Retry
                 return {"foo": "bar"}
@@ -69,7 +81,7 @@ def test_retry_max():
             def start_requests(self):
                 yield Request(server.root_url, callback=self.parse)
 
-            def parse(self, response, page: ItemPage):
+            def parse(self, response, page: SamplePage):
                 items.append(page.to_item())
 
         crawler = make_crawler(TestSpider, {})
@@ -80,23 +92,27 @@ def test_retry_max():
     assert crawler.stats.get_value("retry/count") == 2
     assert crawler.stats.get_value("retry/reason_count/page_object_retry") == 2
     assert crawler.stats.get_value("retry/max_reached") is None
+    _assert_all_unique_instances(page_instances)
+    _assert_all_unique_instances(page_response_instances)
 
 
 @inlineCallbacks
 def test_retry_exceeded():
-    items = []
+    items, page_instances, page_response_instances = [], [], []
 
     with MockServer(EchoResource) as server:
 
-        class ItemPage(WebPage):
+        class SamplePage(WebPage):
             def to_item(self):
+                page_instances.append(self)
+                page_response_instances.append(self.response)
                 raise Retry
 
         class TestSpider(BaseSpider):
             def start_requests(self):
                 yield Request(server.root_url, callback=self.parse)
 
-            def parse(self, response, page: ItemPage):
+            def parse(self, response, page: SamplePage):
                 items.append(page.to_item())
 
         crawler = make_crawler(TestSpider, {})
@@ -107,17 +123,21 @@ def test_retry_exceeded():
     assert crawler.stats.get_value("retry/count") == 2
     assert crawler.stats.get_value("retry/reason_count/page_object_retry") == 2
     assert crawler.stats.get_value("retry/max_reached") == 1
+    _assert_all_unique_instances(page_instances)
+    _assert_all_unique_instances(page_response_instances)
 
 
 @inlineCallbacks
 def test_retry_max_configuration():
     retries = deque([True, True, True, False])
-    items = []
+    items, page_instances, page_response_instances = [], [], []
 
     with MockServer(EchoResource) as server:
 
-        class ItemPage(WebPage):
+        class SamplePage(WebPage):
             def to_item(self):
+                page_instances.append(self)
+                page_response_instances.append(self.response)
                 if retries.popleft():
                     raise Retry
                 return {"foo": "bar"}
@@ -131,7 +151,7 @@ def test_retry_max_configuration():
             def start_requests(self):
                 yield Request(server.root_url, callback=self.parse)
 
-            def parse(self, response, page: ItemPage):
+            def parse(self, response, page: SamplePage):
                 items.append(page.to_item())
 
         crawler = make_crawler(TestSpider, {})
@@ -142,6 +162,52 @@ def test_retry_max_configuration():
     assert crawler.stats.get_value("retry/count") == 3
     assert crawler.stats.get_value("retry/reason_count/page_object_retry") == 3
     assert crawler.stats.get_value("retry/max_reached") is None
+    _assert_all_unique_instances(page_instances)
+    _assert_all_unique_instances(page_response_instances)
+
+
+@inlineCallbacks
+def test_retry_cb_kwargs():
+    retries = deque([True, True, False])
+    items, page_instances, page_response_instances = [], [], []
+
+    with MockServer(EchoResource) as server:
+
+        class SamplePage(WebPage):
+            def to_item(self):
+                page_instances.append(self)
+                page_response_instances.append(self.response)
+                if retries.popleft():
+                    raise Retry
+                return {"foo": "bar"}
+
+        page_from_cb_kwargs = SamplePage(
+            response=HttpResponse("https://example.com", b"")
+        )
+
+        class TestSpider(BaseSpider):
+            def start_requests(self):
+                yield Request(
+                    server.root_url,
+                    callback=self.parse,
+                    cb_kwargs={"page": page_from_cb_kwargs},
+                )
+
+            def parse(self, response, page: SamplePage):
+                items.append(page.to_item())
+
+        crawler = make_crawler(TestSpider, {})
+        yield crawler.crawl()
+
+    assert items == [{"foo": "bar"}]
+    assert crawler.stats.get_value("downloader/request_count") == 3
+    assert crawler.stats.get_value("retry/count") == 2
+    assert crawler.stats.get_value("retry/reason_count/page_object_retry") == 2
+    assert crawler.stats.get_value("retry/max_reached") is None
+    _assert_all_unique_instances(page_instances)
+    _assert_all_unique_instances(page_response_instances)
+    assert page_instances[0] is not page_from_cb_kwargs
+    assert page_response_instances[0] is not page_from_cb_kwargs.response
 
 
 @inlineCallbacks
@@ -150,7 +216,7 @@ def test_non_retry_exception():
 
     with MockServer(EchoResource) as server:
 
-        class ItemPage(WebPage):
+        class SamplePage(WebPage):
             def to_item(self):
                 raise RuntimeError
 
@@ -158,7 +224,7 @@ def test_non_retry_exception():
             def start_requests(self):
                 yield Request(server.root_url, callback=self.parse)
 
-            def parse(self, response, page: ItemPage):
+            def parse(self, response, page: SamplePage):
                 items.append(page.to_item())
 
         crawler = make_crawler(TestSpider, {})